+++ utils/constants.py
@@ -1,0 +1,253 @@
+## Addresses
+FACTORY_ADDRESS = "0xc66F594268041dB60507F00703b152492fb176E7"
+TRIMAKER_ADDRESS = "0xe793455c9728fc91A3E5a33FAfF9eB2F228aE151"
+TRIBAR_ADDRESS = "0x802119e4e253D5C19aA06A5d567C5a41596D6803"
+CHEF_ADDRESS = "0x1f1Ed214bef5E83D8f5d0eB5D7011EB965D0D79B"
+CHEFV2_ADDRESS = "0x3838956710bcc9D122Dd23863a0549ca8D5675D6"
+TRI_ADDRESS = "0xFa94348467f64D5A457F75F8bc40495D33c65aBB"
+WNEAR_ADDRESS = "0xC42C30aC6Cc15faC9bD938618BcaA1a1FaE8501d"
+AURORA_ADDRESS = "0x8BEc47865aDe3B172A928df8f990Bc7f2A3b9f79"
+USDC_ADDRESS = "0xB12BFcA5A55806AaF64E99521918A4bf0fC40802"
+USDT_ADDRESS = "0x4988a896b1227218e4A686fdE5EabdcAbd91571f"
+WETH_ADDRESS = "0xC9BdeEd33CD01541e1eeD10f90519d2C06Fe3feB"
+WBTC_ADDRESS = "0xF4eB217Ba2454613b15dBdea6e5f22276410e89e"
+ATLUNA_ADDRESS = "0xC4bdd27c33ec7daa6fcfd8532ddB524Bf4038096"
+ATUST_ADDRESS = "0x5ce9F0B6AFb36135b5ddBF11705cEB65E634A9dC"
+ASHIBAM_ADDRESS = "0x48687fB162A735a3FedD47a98Fcbf58Be3ed4538"
+FLX_ADDRESS = "0xea62791aa682d455614eaA2A12Ba3d9A2fD197af"
+EMPYR_ADDRESS = "0xE9F226a228Eb58d408FdB94c3ED5A18AF6968fE1"
+AVAX_ADDRESS = "0x80A16016cC4A2E6a2CACA8a4a498b1699fF0f844"
+BNB_ADDRESS = "0x2bF9b864cdc97b08B6D79ad4663e71B8aB65c45c"
+MATIC_ADDRESS = "0x6aB6d61428fde76768D7b45D8BFeec19c6eF91A8"
+ZERO_ADDRESS = "0x0000000000000000000000000000000000000000"
+MECHA_ADDRESS = "0xa33C3B53694419824722C10D99ad7cB16Ea62754"
+META_ADDRESS = "0xc21Ff01229e982d7c8b8691163B0A3Cb8F357453"
+XNL_ADDRESS = "0x7cA1C28663b76CFDe424A9494555B94846205585"
+GBA_ADDRESS = "0xc2ac78FFdDf39e5cD6D83bbD70c1D67517C467eF"
+BBT_ADDRESS = "0x4148d2Ce7816F0AE378d98b40eB3A7211E1fcF0D"
+BSTN_ADDRESS = "0x9f1F933C660a1DC856F0E0Fe058435879c5CCEf0"
+LINEAR_ADDRESS = "0x918dBe087040A41b786f0Da83190c293DAe24749"
+
+### TLP addresses
+WNEAR_USDC = "0x20F8AeFB5697B77E0BB835A8518BE70775cdA1b0"
+WETH_USDC = "0x2F41AF687164062f118297cA10751F4b55478ae1"
+WNEAR_TRI = "0x84b123875F0F36B966d0B6Ca14b31121bd9676AD"
+TRI_AURORA = "0xd1654a7713617d41A8C9530Fb9B948d00e162194"
+MECHA_WNEAR = "0xd62f9ec4C4d323A0C111d5e78b77eA33A2AA862f"
+META_WNEAR = "0xa8CAaf35c0136033294dD286A14051fBf37aed07"
+GBA_USDT = "0x7B273238C6DD0453C160f305df35c350a123E505"
+BBT_WNEAR = "0xadAbA7E2bf88Bd10ACb782302A568294566236dC"
+BSTN_WNEAR = "0xBBf3D4281F10E537d5b13CA80bE22362310b2bf9"
+LINEAR_WNEAR = "0xbceA13f9125b0E3B66e979FedBCbf7A4AfBa6fd1"
+
+## Pool information
+V1_POOLS = {
+    0: "0x63da4DB6Ef4e7C62168aB03982399F9588fCd198",
+    1: "0x20F8AeFB5697B77E0BB835A8518BE70775cdA1b0",
+    2: "0x03B666f3488a7992b2385B12dF7f35156d7b29cD",
+    3: "0x2fe064B6c7D274082aa5d2624709bC9AE7D16C77",
+    4: "0xbc8A244e8fb683ec1Fd6f88F3cc6E565082174Eb",
+    5: "0x84b123875F0F36B966d0B6Ca14b31121bd9676AD",
+    6: "0x5eeC60F348cB1D661E4A5122CF4638c7DB7A886e",
+}
+
+V2_POOLS = {
+        0: {
+            "LP": "0x5eeC60F348cB1D661E4A5122CF4638c7DB7A886e",
+            "LPType": "",
+            "Rewarder": "0x94669d7a170bfe62FAc297061663e0B48C63B9B5",
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": TRI_AURORA,
+            "RewarderTokenDecimals": 18,
+            },
+        1: {
+            "LP": "0xd1654a7713617d41A8C9530Fb9B948d00e162194",
+            "LPType": "",
+            "Rewarder": "0x78EdEeFdF8c3ad827228d07018578E89Cf159Df1",
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": TRI_AURORA,
+            "RewarderTokenDecimals": 18,
+            },
+        2: {
+            "LP": "0xdF8CbF89ad9b7dAFdd3e37acEc539eEcC8c47914",
+            "LPType": "",
+            "Rewarder": "0x89F6628927fdFA2592E016Ba5B14389a4b08D681",
+            "CoingeckoRewarderTokenName": "terra-luna",
+            "RewarderPriceLP": "",
+            "RewarderTokenDecimals": 18,
+            },
+        3: {
+            "LP": "0xa9eded3E339b9cd92bB6DEF5c5379d678131fF90",
+            "LPType": "",
+            "Rewarder": "0x17d1597ec86fD6aecbfE0F32Ab2F2aD9c37E6750",
+            "CoingeckoRewarderTokenName": "terra-luna",
+            "RewarderPriceLP": "",
+            "RewarderTokenDecimals": 18,
+            },
+        4: {
+            "LP": "0x61C9E05d1Cdb1b70856c7a2c53fA9c220830633c",
+            "LPType": "",
+            "Rewarder": ZERO_ADDRESS,
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": "",
+            "RewarderTokenDecimals": "",
+            },
+        5: {
+            "LP": "0x6443532841a5279cb04420E61Cf855cBEb70dc8C",
+            "LPType": "",
+            "Rewarder": ZERO_ADDRESS,
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": "",
+            "RewarderTokenDecimals": "",
+            },
+        6: {
+            "LP": "0x7be4a49AA41B34db70e539d4Ae43c7fBDf839DfA",
+            "LPType": "",
+            "Rewarder": ZERO_ADDRESS,
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": "",
+            "RewarderTokenDecimals": "",
+            },
+        7: {
+            "LP": "0x3dC236Ea01459F57EFc737A12BA3Bb5F3BFfD071",
+            "LPType": "",
+            "Rewarder": ZERO_ADDRESS,
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": "",
+            "RewarderTokenDecimals": "",
+            },
+        8: {
+            "LP": "0x48887cEEA1b8AD328d5254BeF774Be91B90FaA09", 
+            "LPType": "",
+            "Rewarder": "0x42b950FB4dd822ef04C4388450726EFbF1C3CF63",
+            "CoingeckoRewarderTokenName": "flux-token",
+            "RewarderPriceLP": "",
+            "RewarderTokenDecimals": 18,
+            },
+        9: {
+            "LP": "0xd62f9ec4C4d323A0C111d5e78b77eA33A2AA862f", 
+            "LPType": "",
+            "Rewarder": "0x9847F7e33CCbC0542b05d15c5cf3aE2Ae092C057",
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": MECHA_WNEAR,
+            "RewarderTokenDecimals": 18,
+            },
+        10: {
+            "LP": "0xdDAdf88b007B95fEb42DDbd110034C9a8e9746F2",
+            "LPType": "",
+            "Rewarder": "0xbbE41F699B0fB747cd4bA21067F6b27e0698Bc30",
+            "CoingeckoRewarderTokenName": "solace",
+            "RewarderPriceLP": "",
+            "RewarderTokenDecimals": 18,
+            },
+        11: {
+            "LP": "0x5913f644A10d98c79F2e0b609988640187256373",
+            "LPType": "",
+            "Rewarder": "0x7B9e31BbEdbfdc99e3CC8b879b9a3B1e379Ce530",
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": META_WNEAR,
+            "RewarderTokenDecimals": 24,
+            },
+        12: {
+            "LP": "0x47924Ae4968832984F4091EEC537dfF5c38948a4",
+            "LPType": "",
+            "Rewarder": "0xf267212F1D8888e0eD20BbB0c7C87A089cDe6E88",
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": META_WNEAR,
+            "RewarderTokenDecimals": 24,
+            },
+        13: {
+            "LP": "0xb419ff9221039Bdca7bb92A131DD9CF7DEb9b8e5",
+            "LPType": "",
+            "Rewarder": "0xb84293D04137c9061afe34118Dac9931df153826",
+            "CoingeckoRewarderTokenName": "chronicle",
+            "RewarderPriceLP": "",
+            "RewarderTokenDecimals": 18,
+            },
+        14: {
+            "LP": "0xFBc4C42159A5575a772BebA7E3BF91DB508E127a",
+            "LPType": "",
+            "Rewarder": "0x028Fbc4BB5787e340524EF41d95875Ac2C382101",
+            "CoingeckoRewarderTokenName": "chronicle",
+            "RewarderPriceLP": "",
+            "RewarderTokenDecimals": 18,
+            },
+        15: {
+            "LP": "0x7B273238C6DD0453C160f305df35c350a123E505",
+            "LPType": "",
+            "Rewarder": "0xDAc58A615E2A1a94D7fb726a96C273c057997D50",
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": GBA_USDT,
+            "RewarderTokenDecimals": 18,
+            },
+        16: {
+            "LP": "0x6277f94a69Df5df0Bc58b25917B9ECEFBf1b846A",
+            "LPType": "",
+            "Rewarder": "0x170431D69544a1BC97855C6564E8460d39508844",
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": WNEAR_USDC,
+            "RewarderTokenDecimals": 24,
+            },
+        17: {
+            "LP": "0xadAbA7E2bf88Bd10ACb782302A568294566236dC",
+            "LPType": "",
+            "Rewarder": "0xABE01A6b6922130C982E221681EB4C4aD07A21dA",
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": BBT_WNEAR,
+            "RewarderTokenDecimals": 18,
+            },
+        18: {
+            "LP": "0x5EB99863f7eFE88c447Bc9D52AA800421b1de6c9",
+            "LPType": "StableAMM",
+            "Rewarder": ZERO_ADDRESS,
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": "",
+            "RewarderTokenDecimals": "",
+            },
+        19: {
+            "LP": "0x5E74D85311fe2409c341Ce49Ce432BB950D221DE",
+            "LPType": "",
+            "Rewarder": ZERO_ADDRESS,
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": "",
+            "RewarderTokenDecimals": "",
+            },
+        20: {
+            "LP": "0xbe753E99D0dBd12FB39edF9b884eBF3B1B09f26C",
+            "LPType": "",
+            "Rewarder": "0xfe9B7A3bf38cE0CA3D5fA25d371Ff5C6598663d4",
+            "CoingeckoRewarderTokenName": "rose",
+            "RewarderPriceLP": "",
+            "RewarderTokenDecimals": 18,
+            },
+        21: {
+            "LP": "0xbC0e71aE3Ef51ae62103E003A9Be2ffDe8421700",
+            "LPType": "",
+            "Rewarder": "0x87a03aFA70302a5a0F6156eBEd27f230ABF0e69C",
+            "CoingeckoRewarderTokenName": "rose",
+            "RewarderPriceLP": "",
+            "RewarderTokenDecimals": 18,
+            },
+        22: {
+            "LP": "0xbceA13f9125b0E3B66e979FedBCbf7A4AfBa6fd1",
+            "LPType": "",
+            "Rewarder": "0x1616B20534d1d1d731C31Ca325F4e909b8f3E0f0",
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": LINEAR_WNEAR,
+            "RewarderTokenDecimals": 24,
+            },
+        23: {
+            "LP": "0xBBf3D4281F10E537d5b13CA80bE22362310b2bf9",
+            "LPType": "",
+            "Rewarder": "0xDc6d09f5CC085E29972d192cB3AdCDFA6495a741",
+            "CoingeckoRewarderTokenName": "",
+            "RewarderPriceLP": BSTN_WNEAR,
+            "RewarderTokenDecimals": 18,
+            }
+    }
+
+V2_STABLEPOOL_FACTORY = {
+        18: {
+            "poolContract": "0x13e7a001EC72AB30D66E2f386f677e25dCFF5F59"
+        }
+    }
